% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.r
\name{angle_error}
\alias{angle_error}
\title{Angle Error}
\usage{
angle_error(est_cov, ground_truth_cov)
}
\arguments{
\item{est_cov}{Estimated covariance matrix.}

\item{ground_truth_cov}{Ground truth covariance matrix.}
}
\value{
Angle error between eigenvalues.
}
\description{
Computes the angle error between eigenvalues of the estimated covariance matrix and of the ground truth covariance matrix.
}
\details{
The angle error is given by:
\deqn{1-\dfrac{\hat{\boldsymbol{a}}^\top\boldsymbol{a}}{\sqrt{\hat{\boldsymbol{a}}^\top\hat{\boldsymbol{a}}}\sqrt{\boldsymbol{a}^\top\boldsymbol{a}}},}
where \eqn{\hat{\boldsymbol{a}}} and \eqn{\boldsymbol{a}} are the eigenvalues of the estimated and ground truth covariance matrices, respectively.
}
