% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_Shapley.r
\name{barplot_int_Shapley}
\alias{barplot_int_Shapley}
\title{Barplot of Shapley values for Interval-valued Data}
\usage{
barplot_int_Shapley(
  x,
  cutoff_value = NULL,
  cutoff_label = NULL,
  palette = NULL,
  abbrev.var = 20,
  abbrev.obs = 20,
  sort.var = TRUE,
  sort.obs = TRUE,
  plot_IMah = TRUE,
  IMah_label = expression(Robust ~ d[IMah]^2 * (bold(x))),
  rotate_x = TRUE
)
}
\arguments{
\item{x}{A \eqn{n \times p} matrix containing the Shapley values of \eqn{n} observations and \eqn{p} variables.}

\item{cutoff_value}{Numeric. The cutoff value used for detecting outliers. If \code{cutoff_value} is not \code{NULL} (default), the cutoff value is included in the plot.}

\item{cutoff_label}{Character. Label for the cutoff value line in the plot.}

\item{palette}{A vector with colors for each variable. If \code{palette} is \code{NULL} (default), the colors are generated using \code{RColorBrewer}.}

\item{abbrev.var}{Integer. If \code{abbrev.var} \eqn{> 0}, column names are abbreviated using abbreviate with \code{minlenght = abrev.var}.}

\item{abbrev.obs}{Integer. If \code{abbrev.obs} \eqn{> 0}, row names are abbreviated using abbreviate with \code{minlenght = abrev.obs}.}

\item{sort.var}{Logical. If \code{TRUE} (default), variables are sorted according to the  distance.}

\item{sort.obs}{Logical. If \code{TRUE} (default), observations are sorted according to their squared (robust) Interval-Mahalanobis distance.}

\item{plot_IMah}{Logical. If \code{TRUE} (default), the squared (robust) Interval-Mahalanobis distance will be included in the plot.}

\item{IMah_label}{Character. Label for the Interval-Mahalanobis distance in the plot legend. Default is "Robust \eqn{d_\mathrm{IMah}^2(\boldsymbol{x})}".}

\item{rotate_x}{Logical. If \code{TRUE} (default), the x-axis labels are rotated.}
}
\value{
Returns a barplot that displays the Shapley values (\code{\link{int_Shapley}}) for each observation and optionally (\code{plot_IMah = TRUE})
includes the squared (robust) Interval-Mahalanobis distance (\code{\link{IMah_dist}}) (black bar) and the corresponding outlier detection cut-off value (dotted line).
}
\description{
Barplot of Shapley values for Interval-valued Data
}
\examples{
# Create intData object
data(creditcard)
credit_card_int <- creditcard$intData

# Estimate the mean and covariance matrix
credit_card_IMCD <- IMCD(credit_card_int, 
                         m = floor(nrow(credit_card_int)*0.75), 
                         cutoff = "farness", 
                         cutoff_lvl = 0.9)

# Detect outliers using farness cutoff
credit_card_outliers <- int_outliers(credit_card_IMCD$robust_dist, 
                                     cutoff = "farness", 
                                     cutoff_lvl = 0.9)

# Compute Shapley values
credit_card_shapley <- int_Shapley(credit_card_int, 
                                   mean_c = credit_card_IMCD$mean_IMCD_c, 
                                   mean_r = credit_card_IMCD$mean_IMCD_r, 
                                   cov = credit_card_IMCD$cov_IMCD)

# Plot Shapley values with cutoff line and Interval-Mahalanobis distance
barplot_int_Shapley(credit_card_shapley, 
                    cutoff_value = credit_card_outliers$cutoff_value,
                    cutoff_label = "Farness 0.9")
}
\references{
Adapted from package \code{ShapleyOutlier} (\url{https://CRAN.R-project.org/package=ShapleyOutlier}).
}
