% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCD.r
\name{bigIMCD}
\alias{bigIMCD}
\title{Obtain unweighted estimates for data with > 600 observations}
\usage{
bigIMCD(m, p, n, data)
}
\arguments{
\item{m}{An integer specifying number of observations to use}

\item{p}{An integer specifying the number of columns in X}

\item{n}{An integer specifying the number of total observations}

\item{data}{An \code{\linkS4class{intData}} object containing the macrodata/interval data}
}
\value{
A list of z (\code{updated_z}), estimated symbolic covariance (\code{S}), barycenter (\code{mean_c}, \code{mean_r}) and robust distances (\code{robust_dist})
}
\description{
Obtain unweighted estimates for data with > 600 observations
}
\keyword{internal}
