% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_utils.r
\name{cal.E.UU}
\alias{cal.E.UU}
\title{Compute Cal.E Latent Variables}
\usage{
cal.E.UU(
  LatentDist = c("Unif", "Triang", "TNorm", "InvTri", "Beta", "KDE", "Degenerated"),
  TriangParam = 0,
  BetaParam.a = 1,
  BetaParam.b = 1,
  Umicro = NULL,
  p = NULL
)
}
\arguments{
\item{LatentDist}{A string or vector of strings specifying the distribution(s) of the latent variables. If the variables are identically distributed it can be one of (\code{"Unif"},\code{"Triang"},\code{"TNorm"},\code{"InvTri"},\code{"Beta"},\code{"KDE"},\code{"Degenerated"}), if not a vector must be provided with the distribution for each variable.}

\item{TriangParam}{Mode of the triangular distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{0}.}

\item{BetaParam.a}{Parameter alpha of the Beta distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{1}.}

\item{BetaParam.b}{Parameter beta of the Beta distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{1}.}

\item{Umicro}{Latent microdata observations. Needed if \code{LatentDist="KDE"}.}

\item{p}{Number of variables.}
}
\value{
A \eqn{p\times p} matrix.
}
\description{
Computes \eqn{\boldsymbol{\mathfrak{E}}_{UU}} for the latent variables inherent to the macrodata.
}
\details{
The matrix \eqn{\boldsymbol{\mathfrak{E}}_{UU}} is defined as follows:
\itemize{
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} , \eqn{i\neq j}, with \eqn{\mathcal{E}(U_i,U_j)=\int_0^1 F_{U_i}^{-1}(t) F_{U_j}^{-1}(t) \, dt}
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ii}=\mathbb{E}(U_i^2)}, \eqn{i,j=1,\dots,p}.
}
}
\keyword{internal}
