% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_utils.r
\name{CalE.beta.beta}
\alias{CalE.beta.beta}
\title{Computes \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} for the latent variables inherent to the macrodata, where they follow a Beta distribution.}
\usage{
CalE.beta.beta(a1, b1, a2, b2)
}
\arguments{
\item{a1}{Parameter alpha of the first Beta distribution.}

\item{b1}{Parameter beta of the first Beta distribution.}

\item{a2}{Parameter alpha of the second Beta distribution.}

\item{b2}{Parameter beta of the second Beta distribution.}
}
\value{
Value
}
\description{
Computes \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} for the latent variables inherent to the macrodata, where they follow a Beta distribution.
}
\keyword{internal}
