% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_utils.r
\name{CalE.beta.kde}
\alias{CalE.beta.kde}
\title{Computes \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} for the latent variables inherent to the macrodata, where U_1 follows a Beta(a_1,b_1) and the PDF of U_2 is estimated by a KDE.}
\usage{
CalE.beta.kde(micro, a1, b1)
}
\arguments{
\item{micro}{Latent microdata observations.}

\item{a1}{Parameter alpha of the Beta distribution.}

\item{b1}{Parameter beta of the Beta distribution.}
}
\value{
Value
}
\description{
Computes \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} for the latent variables inherent to the macrodata, where U_1 follows a Beta(a_1,b_1) and the PDF of U_2 is estimated by a KDE.
}
\keyword{internal}
