% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_utils.r
\name{CalE.kde.kde}
\alias{CalE.kde.kde}
\title{Computes \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} for the latent variables inherent to the macrodata, where the PDF is estimated by a KDE.}
\usage{
CalE.kde.kde(micro1, micro2)
}
\arguments{
\item{micro1}{Latent microdata observations of the first latent variable.}

\item{micro2}{Latent microdata observations of the second latent variable.}
}
\value{
Value
}
\description{
Computes \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} for the latent variables inherent to the macrodata, where the PDF is estimated by a KDE.
}
\keyword{internal}
