% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_utils.r
\name{CalE.triang.triang}
\alias{CalE.triang.triang}
\title{Computes \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} for the latent variables inherent to the macrodata, where they follow a Triangular distribution.}
\usage{
CalE.triang.triang(mo1 = 0, mo2 = 0)
}
\arguments{
\item{mo1}{Mode of the triangular distribution of the first latent variable.}

\item{mo2}{Mode of the triangular distribution of the second latent variable.}
}
\value{
Value
}
\description{
Computes \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)} for the latent variables inherent to the macrodata, where they follow a Triangular distribution.
}
\keyword{internal}
