% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{Centers}
\alias{Centers}
\alias{Centers,intData-method}
\title{Centers Method for \linkS4class{intData}}
\usage{
Centers(Sdt)

\S4method{Centers}{intData}(Sdt)
}
\arguments{
\item{Sdt}{An object of class \linkS4class{intData}.}
}
\value{
A \code{data.frame} containing the centers of the intervals.
}
\description{
Centers Method for \linkS4class{intData}
}
