% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{colnames,intData-method}
\alias{colnames,intData-method}
\title{Column Names Method for \linkS4class{intData}}
\usage{
\S4method{colnames}{intData}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{intData}.}
}
\value{
A character vector of column names.
}
\description{
Column Names Method for \linkS4class{intData}
}
