% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{==,intData,intData-method}
\alias{==,intData,intData-method}
\alias{!=,intData,intData-method}
\title{Equality Comparison for \code{\linkS4class{intData}} Objects}
\usage{
\S4method{==}{intData,intData}(e1, e2)

\S4method{!=}{intData,intData}(e1, e2)
}
\arguments{
\item{e1}{An \code{\linkS4class{intData}} object.}

\item{e2}{An \code{\linkS4class{intData}} object.}
}
\value{
A logical matrix indicating which elements are equal between the two \code{\linkS4class{intData}} objects.

A logical matrix indicating element-wise inequality of the two \code{\linkS4class{intData}} objects.
}
\description{
Compare two \code{\linkS4class{intData}} objects for equality.

Compare two \code{\linkS4class{intData}} objects for inequality.
}
