% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{dim,intData-method}
\alias{dim,intData-method}
\title{Dimensions Method for \linkS4class{intData}}
\usage{
\S4method{dim}{intData}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{intData}.}
}
\value{
A vector of the number of rows and columns.
}
\description{
Dimensions Method for \linkS4class{intData}
}
