% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCD.r
\name{draw_z}
\alias{draw_z}
\title{Randomly draw a subset of observations}
\usage{
draw_z(m, data)
}
\arguments{
\item{m}{An integer specifying the number of observations to use}

\item{data}{An \linkS4class{intData} object containing the macrodata/interval data}
}
\value{
A vector representing an m-length subset of X
}
\description{
Randomly draw a subset of observations
}
\keyword{internal}
