% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.r
\name{frobenius_error}
\alias{frobenius_error}
\title{Relative Frobenius Error}
\usage{
frobenius_error(est_cov, ground_truth_cov)
}
\arguments{
\item{est_cov}{Estimated covariance matrix.}

\item{ground_truth_cov}{Ground truth covariance matrix.}
}
\value{
Frobenius error between the two matrices.
}
\description{
Computes the relative Frobenius error between an estimated covariance matrix and the ground truth.
}
\details{
The relative Frobenius error is given by:
\deqn{\dfrac{\|\boldsymbol{A} - \boldsymbol{B}\|_F}{\|\boldsymbol{B}\|_F}=\dfrac{\sqrt{\sum\limits_{i=1}^{p}\sum\limits_{j=1}^{p}|[\boldsymbol{A}]_{ij}-[\boldsymbol{B}]_{ij}|^2}}{\sqrt{\sum\limits_{i=1}^{p}\sum\limits_{j=1}^{p}|[\boldsymbol{B}]_{ij}|^2}},}
where \eqn{\boldsymbol{A}} and \eqn{\boldsymbol{B}} are the estimated and ground truth covariance matrices, respectively.
}
