% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_utils.r
\name{get_latent_param}
\alias{get_latent_param}
\title{Compute Latent Variables Parameters}
\usage{
get_latent_param(
  LatentCase = c("U_id_symmetric", "U_id", "General"),
  LatentDist = c("Unif", "Triang", "TNorm", "InvTri", "Beta", "KDE", "Degenerated"),
  TriangParam = 0,
  BetaParam.a = 1,
  BetaParam.b = 1,
  Umicro = NULL,
  p = NULL,
  estimate.DistParam = FALSE
)
}
\arguments{
\item{LatentCase}{A string specifying which of the three scenarios applies to the latent variables:
\itemize{
\item \code{"U_id_symmetric"}: The case where the latent variables are identically distributed and symmetric.
\item \code{"U_id"}: The case where the latent variables are identically distributed.
\item \code{"General"}: The case where the latent variables do not have any nice properties.
}
Defaults to \code{"U_id_symmetric"}.}

\item{LatentDist}{A string or vector of strings specifying the distribution(s) of the latent variables. If the variables are identically distributed it can be one of (\code{"Unif"}, \code{"Triang"}, \code{"TNorm"}, \code{"InvTri"}, \code{"Beta"}, \code{"KDE"}, \code{"Degenerated"}), if not a vector must be provided with the distribution for each variable.
The default is \code{"Unif"} if \code{LatentCase="U_id_symmetric"} or if \code{Umicro} is not provided, and \code{"KDE"} if \code{LatentCase="General"}.}

\item{TriangParam}{Mode of the triangular distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{0}.}

\item{BetaParam.a}{Parameter alpha of the Beta distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{1}.}

\item{BetaParam.b}{Parameter beta of the Beta distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{1}.}

\item{Umicro}{Latent microdata observations.
Needed if \code{estimate.DistParam} is \code{TRUE} or \code{LatentDist} is \code{"KDE"}.}

\item{p}{Number of variables.}

\item{estimate.DistParam}{Logical parameter indicating if estimation of the parameters of the latent distributions should be performed. Can only be set to TRUE if \code{LatentCase="General"}.
The default is \code{FALSE}.}
}
\value{
A list composed by:
\item{\code{LatentParam}}{A list with the parameters of the latent variables distribution.}
\item{\code{TriangParam}}{The mode of the triangular distribution.}
\item{\code{BetaParam.a}}{The alpha parameter of the Beta distribution.}
\item{\code{BetaParam.b}}{The beta parameter of the Beta distribution.}
\item{\code{LatentCase}}{The case of the latent variables.}
\item{\code{LatentDist}}{The distribution of the latent variables.}
}
\description{
Obtain the parameters of the latent variables inherent to the macrodata.
}
\details{
The parameters of the latent variables inherent to the macrodata are defined according to the \code{LatentCase}:
\itemize{
\item \code{"U_id_symmetric"}: The latent variables are identically distributed and symmetric, so its parameters are:
\itemize{
\item  \eqn{\delta=\mathbb{E}(U^2)/4}
}
\item \code{"U_id"}: The latent variables are identically distributed, so its parameters are:
\itemize{
\item  \eqn{\delta=\mathbb{E}(U^2)/4}
\item  \eqn{\mathbb{E}(U)}
}
\item \code{"General"}: The latent variables do not have any nice properties, so its parameters are:
\itemize{
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{j\ell}=\mathcal{E}(U_j,U_\ell)}, \eqn{j\neq \ell}, with \eqn{\mathcal{E}(U_j,U_\ell)=\int_0^1 F_{U_j}^{-1}(t) F_{U_\ell}^{-1}(t) \, dt}, and \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{jj}=\mathbb{E}(U_j^2)}, \eqn{j,\ell=1,\dots,p}
\item \eqn{\boldsymbol{\Psi}=\text{Diag}(\mathbb{E}(U_1),\dots,\mathbb{E}(U_p))}
}
}
}
\examples{
data(creditcard)
CreditCard_min_max <- creditcard$min_max
CreditCard_microdata <- creditcard$microdata

# Define grouping variable for microdata aggregation
credit_agrby <- paste(CreditCard_microdata$Name, CreditCard_microdata$Month, sep = "_")

# Obtain latent variables inherent to the macrodata (normalized to [-1,1])
credit_card_U <- get_latent_var(microdata = CreditCard_microdata[,3:7], 
                                macrodata = CreditCard_min_max, 
                                agrby = credit_agrby, 
                                agrlevels = row.names(CreditCard_min_max), 
                                Seq = "LbUb_VarbyVar")

# Obtain parameters of the latent variables
credit_card_param <- get_latent_param(LatentCase = "General",
                                      LatentDist = "KDE",
                                      Umicro = credit_card_U)
}
\references{
Oliveira, M. R., Pinheiro, D., & Oliveira, L. (2025).
Location and association measures for interval-valued data based on Mallows' distance.
arXiv preprint arXiv:2407.05105. \url{https://arxiv.org/abs/2407.05105}
}
