% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{head,intData-method}
\alias{head,intData-method}
\title{Head Method for \linkS4class{intData}}
\usage{
\S4method{head}{intData}(x, n = min(nrow(x), 6L))
}
\arguments{
\item{x}{An \linkS4class{intData} object.}

\item{n}{The number of rows to return.}
}
\value{
A subset of the \linkS4class{intData} object.
}
\description{
Returns the first \code{n} rows of an \linkS4class{intData} object.
}
