% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_distances.r
\name{IMah_dist}
\alias{IMah_dist}
\title{Interval-Mahalanobis Distance}
\usage{
IMah_dist(data, z = NULL, mean_c = NULL, mean_r = NULL, cov = NULL)
}
\arguments{
\item{data}{An \code{\linkS4class{intData}} object containing the macrodata/interval data}

\item{z}{(Optional) A vector of 0 and 1, indicating which observations should be considered for the calculation.
If \code{z} is not \code{NULL}, \code{mean_c}, \code{mean_r}, and \code{cov} will be computed using only the observations with \code{z=1} (see \code{\link{int_mean_z}} and \code{\link{int_cov_z}}).
Defaults to \code{NULL}.}

\item{mean_c}{(Optional) A vector specifying the mean of centers. Defaults to \code{NULL}, in which case
it will be computed using the \code{\link{IMCD}} function, if \code{z} is also \code{NULL}.}

\item{mean_r}{(Optional) A vector specifying the mean of ranges. Defaults to \code{NULL}, in which case
it will be computed using the \code{\link{IMCD}} function, if \code{z} is also \code{NULL}.}

\item{cov}{(Optional) A covariance matrix. Defaults to \code{NULL}, in which case it will be computed
using the \code{\link{IMCD}} function, if \code{z} is also \code{NULL}.}
}
\value{
A vector with the squared Interval-Mahalanobis distance of each observation.
}
\description{
Calculate the squared Interval-Mahalanobis distance of all rows in the data and the barycenter.
}
\details{
The squared Interval-Mahalanobis distance between \eqn{\boldsymbol{x}=(\boldsymbol{c}^\top,\boldsymbol{r}^\top)^\top} and the barycenter \eqn{\boldsymbol{\mu}_B=(\boldsymbol{\mu}_C^\top,\boldsymbol{\mu}_R^\top)^\top} of a population with symbolic covariance matrix \eqn{\boldsymbol{\Sigma}_B} (see \code{\link{int_cov}}) is defined according to the \code{LatentCase}:
\itemize{
\item \code{"U_id_symmetric"}: The latent variables are identically distributed and symmetric:
\deqn{d_\mathrm{IMah}(\boldsymbol{x})^2=(\boldsymbol{c}-\boldsymbol{\mu}_C)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{c}-\boldsymbol{\mu}_C)+\delta(\boldsymbol{r}-\boldsymbol{\mu}_R)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{r}-\boldsymbol{\mu}_R),}
where \eqn{\delta=\mathbb{E}(U^2)/4} is the parameter of the latent variables.
\item \code{"U_id"}: The latent variables are identically distributed:
\deqn{\begin{aligned}
             d_\mathrm{IMah}(\boldsymbol{x})^2&=(\boldsymbol{c}-\boldsymbol{\mu}_C)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{c}-\boldsymbol{\mu}_C)+\delta(\boldsymbol{r}-\boldsymbol{\mu}_R)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{r}-\boldsymbol{\mu}_R)\\
             &\quad+\mathbb{E}(U)(\boldsymbol{c}-\boldsymbol{\mu}_C)^\top\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{r}-\boldsymbol{\mu}_R),
    \end{aligned}}
where \eqn{\delta=\mathbb{E}(U^2)/4} and \eqn{\mathbb{E}(U)} are the parameter of the latent variables.
\item \code{"General"}: The latent variables do not have any nice properties:
\deqn{\begin{aligned}
             d_\mathrm{IMah}(\boldsymbol{x})^2&=(\boldsymbol{c}-\boldsymbol{\mu}_C)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{c}-\boldsymbol{\mu}_C)+\dfrac{1}{4}(\boldsymbol{r}-\boldsymbol{\mu}_R)^{\top}\left(\boldsymbol{\mathfrak{E}}_{UU}\bullet\boldsymbol{\Sigma}_{B}^{-1}\right)(\boldsymbol{r}-\boldsymbol{\mu}_R)\\
             &\quad+(\boldsymbol{c}-\boldsymbol{\mu}_C)^{\top}\boldsymbol{\Sigma}_{B}^{-1}\boldsymbol{\Psi}(\boldsymbol{r}-\boldsymbol{\mu}_R),
   \end{aligned}}
where:
\itemize{
\item \eqn{\boldsymbol{\Psi}=\text{diag}(\mathbb{E}(U_1),\dots,\mathbb{E}(U_p))},
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{j\ell}=\mathcal{E}(U_j,U_\ell)}, \eqn{j\neq \ell}, with \eqn{\mathcal{E}(U_j,U_\ell)=\int_0^1 F_{U_j}^{-1}(t) F_{U_\ell}^{-1}(t) \, dt},
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{jj}=\mathbb{E}(U_j^2)}, \eqn{j,\ell=1,\dots,p},
\item \eqn{\bullet} denotes the Schur (or entrywise) product of matrices.
}
}
}
\examples{
data(creditcard)
credit_card_int <- creditcard$intData

# Compute squared Interval-Mahalanobis distance using IMCD estimates of mean and covariance
credit_card_dist <- IMah_dist(credit_card_int)
}
\references{
Loureiro, C. P., Oliveira, M. R., Brito, P., & Oliveira, L. (2026).
Minimum Covariance Determinant Estimator and Outlier Detection for Interval-valued Data.
arXiv preprint arXiv:2604.26769. \url{https://arxiv.org/abs/2604.26769}
}
