% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_cov.r
\name{int_cov}
\alias{int_cov}
\title{Interval-valued Covariance}
\usage{
int_cov(
  data = NULL,
  sigma_cc = NULL,
  sigma_rr = NULL,
  sigma_cr = NULL,
  LatentParam = NULL,
  LatentCase = c("U_id_symmetric", "U_id", "General")
)
}
\arguments{
\item{data}{An \code{\linkS4class{intData}} object containing the macrodata/interval data.
If \code{data} is provided, the covariance matrix is calculated based on the the sample covariance of the centers and ranges and the sample covariance between centers and ranges, and the parameters of the latent variables contained in the \code{intData} object.
If \code{data} is not provided, the covariance matrix is calculated based on \code{sigma_cc}, \code{sigma_rr}, \code{sigma_cr}, \code{LatentParam}, and \code{LatentCase}.}

\item{sigma_cc}{Covariance matrix of the centers.}

\item{sigma_rr}{Covariance matrix of the ranges.}

\item{sigma_cr}{Covariance matrix between the centers and ranges.}

\item{LatentParam}{A list with the parameters of the latent variables.
Expects a list with a single number if \code{LatentCase} is \code{"U_id_symmetric"}, a list of two numbers if \code{LatentCase} is \code{"U_id"}, and a list of two matrices if \code{LatentCase} is \code{"General"}.}

\item{LatentCase}{A string specifying which of the three scenarios applies to the latent variables:
\itemize{
\item \code{"U_id_symmetric"}: The case where the latent variables are identically distributed and symmetric.
\item \code{"U_id"}: The case where the latent variables are identically distributed.
\item \code{"General"}: The case where the latent variables do not have any nice properties.
}
Defaults to \code{"U_id_symmetric"}.}
}
\value{
The symbolic covariance matrix.
}
\description{
Calculate the interval-valued covariance matrix based on the covariance matrices of the centers and ranges or data.
}
\details{
This function calculates the interval-valued covariance matrix, \eqn{\boldsymbol{\Sigma}_B}, based on the covariance matrices of the centers, \eqn{\boldsymbol{\Sigma}_{CC}}, ranges, \eqn{\boldsymbol{\Sigma}_{RR}}, and the covariance matrix between the centers and ranges, \eqn{\boldsymbol{\Sigma}_{CR}=\boldsymbol{\Sigma}_{RC}^\top}.
The covariance matrix is defined according to the \code{LatentCase}:
\itemize{
\item \code{"U_id_symmetric"}: The latent variables are identically distributed and symmetric:
\deqn{\boldsymbol{\Sigma}_B=\boldsymbol{\Sigma}_{CC}+\delta\boldsymbol{\Sigma}_{RR},}
where \eqn{\delta=\mathbb{E}(U^2)/4} is the parameter of the latent variables.
\item \code{"U_id"}: The latent variables are identically distributed:
\deqn{\boldsymbol{\Sigma}_B=\boldsymbol{\Sigma}_{CC}+\delta\boldsymbol{\Sigma}_{RR}+\dfrac{\mathbb{E}(U)}{2}\left(\boldsymbol{\Sigma}_{CR}+\boldsymbol{\Sigma}_{RC}\right),}
where \eqn{\delta=\mathbb{E}(U^2)/4} and \eqn{\mathbb{E}(U)} are the parameters of the latent variables.
\item \code{"General"}: The latent variables do not have any nice properties:
\deqn{\boldsymbol{\Sigma}_B=\boldsymbol{\Sigma}_{CC}+\dfrac{1}{4}\boldsymbol{\mathfrak{E}}_{UU}\bullet\boldsymbol{\Sigma}_{RR}+\dfrac{1}{2}\boldsymbol{\Sigma}_{CR}\boldsymbol{\Psi}+\dfrac{1}{2}\boldsymbol{\Psi}\boldsymbol{\Sigma}_{RC}}
where:
\itemize{
\item \eqn{\boldsymbol{\Psi}=\text{diag}(\mathbb{E}(U_1),\dots,\mathbb{E}(U_p))},
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{j\ell}=\mathcal{E}(U_j,U_\ell)}, \eqn{j\neq \ell}, with \eqn{\mathcal{E}(U_j,U_\ell)=\int_0^1 F_{U_j}^{-1}(t) F_{U_\ell}^{-1}(t) \, dt},
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{jj}=\mathbb{E}(U_j^2)}, \eqn{j,\ell=1,\dots,p},
\item \eqn{\bullet} denotes the Schur (or entrywise) product of matrices.
}
}
The covariance matrix can be calculated either based on the covariance matrices of the centers and ranges or based on the data. If the data is provided, the covariance matrices are calculated using the sample covariance of the centers and ranges and the sample covariance between centers and ranges.
For the robust estimation of the covariance matrix, see \code{\link{IMCD}}.
}
\examples{
data(creditcard)
credit_card_int <- creditcard$intData

credit_card_cov <- int_cov(credit_card_int)
}
\references{
Oliveira, M. R., Pinheiro, D., & Oliveira, L. (2025).
Location and association measures for interval-valued data based on Mallows' distance.
arXiv preprint arXiv:2407.05105. \url{https://arxiv.org/abs/2407.05105}
}
