% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_cov.r
\name{int_mean_z}
\alias{int_mean_z}
\title{Sample Mean}
\usage{
int_mean_z(z, X)
}
\arguments{
\item{z}{A vector of 0 and 1, indicating which observations should be considered for the calculation}

\item{X}{A matrix where the rows correspond to observations and the columns to variables}
}
\value{
A vector where each element is the mean for each variable
}
\description{
Calculate the mean of X in function of z
}
\details{
This function calculates the mean of \eqn{\boldsymbol{X}} in function of \eqn{\boldsymbol{z}}. If \eqn{\boldsymbol{z}} is a vector of 0 and 1, the mean is calculated for the \eqn{m} observations that are equal to 1:
\deqn{\bar{\boldsymbol{x}}(\boldsymbol{z}) = \dfrac{1}{m} \boldsymbol{X}^\top \boldsymbol{z}.}
}
\examples{
n <- 100
p <- 4
X <- matrix(rnorm(n * p), ncol = p)
#if we consider all the observations the result obtained is the same as colMeans()
z <- c(rep(1, n))
int_mean_z(z, X)
colMeans(X)
}
