% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_outliers.r
\name{int_outliers}
\alias{int_outliers}
\title{Outlier Detection for Interval-Valued Data Based on Robust Distances}
\usage{
int_outliers(
  robust_dist,
  cutoff = c("farness", "adjbox", "chi-squared", "F-dist"),
  cutoff_lvl = NULL,
  p = NULL,
  z = NULL
)
}
\arguments{
\item{robust_dist}{A numeric vector containing the robust distances for each observation.}

\item{cutoff}{A character string specifying the method for setting the outlier cutoff threshold. Options include:
\itemize{
\item \code{"chi-squared"}: Outliers are identified based on a specified Chi-Squared quantile.
\item \code{"adjbox"}: Uses adjusted boxplot statistics (from \code{robustbase}) to classify outliers.
\item \code{"F-dist"}: Applies a cutoff derived from the F and Beta distributions for robust outlier detection.
\item \code{"farness"}: Identifies outliers based on a "farness" threshold, determined by the robust distance distribution.
}
Default is \code{"farness"}.}

\item{cutoff_lvl}{A numeric value specifying the level of the cutoff to be used.
\itemize{
\item If \code{cutoff="chi-squared"}, \code{cutoff_lvl} is the quantile of the Chi-squared distribution (default is 0.975).
\item If \code{cutoff="adjbox"}, \code{cutoff_lvl} is the coefficient for the adjusted boxplot (default is 1.5).
\item If \code{cutoff="F-dist"}, \code{cutoff_lvl} is the significance level for identifying outliers (default is 0.95).
\item If \code{cutoff="farness"}, \code{cutoff_lvl} represents the threshold for farness, with a default of 0.99.
}
If no value is provided, the function uses the default values associated with each cutoff method.}

\item{p}{The number of variables in the data. Required for \code{"chi-squared"} and \code{"F-dist"} cutoff methods.}

\item{z}{A binary vector indicating the subset of observations used for initial robust estimation. Required for the \code{"F-dist"} cutoff method.}
}
\value{
A list with the following components:
\item{\code{outliers_names}}{Character vector of names for observations classified as outliers.}
\item{\code{is_outlier}}{Logical vector indicating whether each observation is an outlier (TRUE) or not (FALSE).}
\item{\code{cutoff}}{The cutoff method used for detecting outliers.}
\item{\code{cutoff_value}}{Cutoff value used for detecting outliers.}
\item{\code{farness_probs}}{Numeric vector of farness probabilities for each observation (only if \code{cutoff} is set to \code{"farness"}).}
}
\description{
Identifies potential outliers in interval-valued data using robust distance-based methods with customizable cutoff criteria.
}
\details{
This function classifies observations as outliers based on robust distances and user-defined cutoff methods. It supports various approaches, including Chi-Squared quantiles, adjusted boxplots, F distribution quantiles, and farness probabilities.
}
\examples{
# Example of detecting outliers using robust distances
set.seed(42)
robust_dist <- abs(rnorm(100))
result <- int_outliers(robust_dist, cutoff="chi-squared", p=5)

# Example using creditcard dataset
data(creditcard)
credit_card_int <- creditcard$intData

credit_card_IMCD <- IMCD(credit_card_int, floor(0.75*credit_card_int@NObs), "farness", 0.9)
credit_card_outliers <- int_outliers(credit_card_IMCD$robust_dist, "farness", 0.9)
}
\references{
Loureiro, C. P., Oliveira, M. R., Brito, P., & Oliveira, L. (2026).
Minimum Covariance Determinant Estimator and Outlier Detection for Interval-valued Data.
arXiv preprint arXiv:2604.26769. \url{https://arxiv.org/abs/2604.26769}

Case \code{cutoff=="F-dist"} is adapted from package \code{CerioliOutlierDetection} (\url{https://cran.r-project.org/package=CerioliOutlierDetection}).
}
