% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.r
\name{int_Shapley_interaction}
\alias{int_Shapley_interaction}
\title{Compute Shapley interaction indices for Interval-valued Data}
\usage{
int_Shapley_interaction(data, mean_c = NULL, mean_r = NULL, cov = NULL)
}
\arguments{
\item{data}{An \code{\linkS4class{intData}} object containing the interval-valued dataset (macrodata).}

\item{mean_c}{(Optional) A vector specifying the mean of centers. Defaults to \code{NULL}, in which case
it will be computed using the \code{\link{IMCD}} function.}

\item{mean_r}{(Optional) A vector specifying the mean of ranges. Defaults to \code{NULL}, in which case
it will be computed using the \code{\link{IMCD}} function.}

\item{cov}{(Optional) A covariance matrix. Defaults to \code{NULL}, in which case it will be computed
using the \code{\link{IMCD}} function.}
}
\value{
A list containing the matrix of Shapley interaction indices for each observation.
}
\description{
Obtains a \eqn{p \times p} matrix containing pairwise outlyingness scores based on Shapley interaction indices for each observation.
Decomposes the squared interval-valued Mahalanobis distance of each observation into outlyingness contributions of pairs of variables.
}
\details{
Let \eqn{\boldsymbol{\mu}_B=(\boldsymbol{\mu}_C^\top,\boldsymbol{\mu}_R^\top)^\top} be the barycenter and \eqn{\boldsymbol{\Sigma}_B} the symbolic covariance matrix (see \code{\link{int_cov}}).
Let also \eqn{\boldsymbol{\phi}(\boldsymbol{x})} be the Shapley value of \eqn{\boldsymbol{x}} (see \code{\link{int_Shapley}}) and \eqn{\mathrm{diag}(\boldsymbol{v})} be the diagonal matrix whose main diagonal is the vector \eqn{\boldsymbol{v}}.
The Shapley interaction index of an interval-valued observation \eqn{\boldsymbol{x}=(\boldsymbol{c}^\top,\boldsymbol{r}^\top)^\top}, for the Interval-Mahalanobis distance, is defined according to the \code{LatentCase}:
\itemize{
\item \code{"U_id_symmetric"}: The latent variables are identically distributed and symmetric:
\deqn{\boldsymbol{\Phi}(\boldsymbol{x})=2(\boldsymbol{c}-\boldsymbol{\mu}_C)(\boldsymbol{c}-\boldsymbol{\mu}_C)^\top\bullet\boldsymbol{\Sigma}_B^{-1} + 2\delta(\boldsymbol{r}-\boldsymbol{\mu}_R)(\boldsymbol{r}-\boldsymbol{\mu}_R)^\top\bullet\boldsymbol{\Sigma}_B^{-1}-\mathrm{diag}\left(\boldsymbol{\phi}(\boldsymbol{x})\right),}
where \eqn{\delta=\mathbb{E}(U^2)/4} is the parameter of the latent variables.
\item \code{"U_id"}: The latent variables are identically distributed:
\deqn{\begin{aligned}
            \boldsymbol{\Phi}(\boldsymbol{x})&=2(\boldsymbol{c}-\boldsymbol{\mu}_C)(\boldsymbol{c}-\boldsymbol{\mu}_C)^\top\bullet\boldsymbol{\Sigma}_B^{-1} + 2\delta(\boldsymbol{r}-\boldsymbol{\mu}_R)(\boldsymbol{r}-\boldsymbol{\mu}_R)^\top\bullet\boldsymbol{\Sigma}_B^{-1}\\
           &\quad+\mathbb{E}(U)(\boldsymbol{c}-\boldsymbol{\mu}_C)(\boldsymbol{r}-\boldsymbol{\mu}_R)^\top\bullet\boldsymbol{\Psi} + \mathbb{E}(U)(\boldsymbol{r}-\boldsymbol{\mu}_R)(\boldsymbol{c}-\boldsymbol{\mu}_C)^\top\bullet\boldsymbol{\Sigma}_B^{-1}-\mathrm{diag}\left(\boldsymbol{\phi}(\boldsymbol{x})\right),
   \end{aligned}}
where \eqn{\delta=\mathbb{E}(U^2)/4} and \eqn{\mathbb{E}(U)} are the parameter of the latent variables.
\item \code{"General"}: The latent variables do not have any nice properties:
\deqn{\begin{aligned}
      \boldsymbol{\Phi}(\boldsymbol{x})&=2(\boldsymbol{c}-\boldsymbol{\mu}_C)(\boldsymbol{c}-\boldsymbol{\mu}_C)^\top\bullet\boldsymbol{\Sigma}_B^{-1} + \dfrac{1}{2}(\boldsymbol{r}-\boldsymbol{\mu}_R)(\boldsymbol{r}-\boldsymbol{\mu}_R)^\top\bullet\boldsymbol{\mathfrak{E}}_{UU}\bullet\boldsymbol{\Sigma}_B^{-1}\\
     &\quad+(\boldsymbol{c}-\boldsymbol{\mu}_C)(\boldsymbol{r}-\boldsymbol{\mu}_R)^\top\bullet\boldsymbol{\Sigma}_B^{-1}\boldsymbol{\Psi} + (\boldsymbol{r}-\boldsymbol{\mu}_R)(\boldsymbol{c}-\boldsymbol{\mu}_C)^\top\bullet\boldsymbol{\Psi}\boldsymbol{\Sigma}_B^{-1}-\mathrm{diag}\left(\boldsymbol{\phi}(\boldsymbol{x})\right),
  \end{aligned}}
where:
\item \eqn{\boldsymbol{\Psi}=\text{Diag}(\mathbb{E}(U_1),\dots,\mathbb{E}(U_p))},
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{j\ell}=\mathcal{E}(U_j,U_\ell)}, \eqn{j\neq \ell}, with \eqn{\mathcal{E}(U_j,U_\ell)=\int_0^1 F_{U_j}^{-1}(t) F_{U_\ell}^{-1}(t) \, dt},
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{jj}=\mathbb{E}(U_j^2)}, \eqn{j,\ell=1,\dots,p},
\item \eqn{\bullet} denotes the Schur (or entrywise) product of matrices.
}
}
\examples{
# Create intData object
data(creditcard)
credit_card_int <- creditcard$intData

# Compute Shapley interaction indices based on the mean and covariance matrix estimated by IMCD
credit_card_shap_inter <- int_Shapley_interaction(credit_card_int)
}
\references{
Loureiro, C. P., Oliveira, M. R., Brito, P., & Oliveira, L. (2026).
Explainable Outlier Detection for Interval-valued Data.
}
