% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{intCars}
\alias{intCars}
\title{Cars Dataset}
\format{
A list with the following components:
\describe{
\item{\code{min_max}}{A data frame with \code{27} rows and \code{9} columns. It contains the lower and upper bounds for each variable.}
\item{\code{intData}}{An \code{\linkS4class{intData}} object with \code{27} interval-valued observations and \code{4} variables. The variable "Price" was log-transformed into "lnPrice". The microdata are not available, thus the default parameters of the latent distributions were used assuming a uniform distribution.}
}
}
\usage{
data(intCars)
}
\description{
This dataset contains interval data of car specifications, including min-max values.
It is composed of 5 variables: Engine Capacity, Top Speed, Acceleration, Price and Class.
The aggregation of the microdata was done by car model.
}
\examples{
data(intCars)
head(intCars$min_max)
head(intCars$intData)

}
\references{
This data was retrieved from the \code{MAINT.Data} package, available at \url{https://cran.r-project.org/package=MAINT.Data}.
}
\keyword{datasets}
