% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.r
\name{KL_divergence}
\alias{KL_divergence}
\title{Kullback-Leibler (KL) Divergence}
\usage{
KL_divergence(est_cov, ground_truth_cov)
}
\arguments{
\item{est_cov}{Estimated covariance matrix.}

\item{ground_truth_cov}{Ground truth covariance matrix.}
}
\value{
KL divergence between the two matrices.
}
\description{
Computes the Kullback-Leibler (KL) divergence between an estimated covariance matrix and the ground truth. Assumes normal multivariate distributions.
}
\details{
The KL divergence between two \eqn{p}-dimensional Gaussians \eqn{\mathcal{N}(\boldsymbol{\mu}, \hat{\boldsymbol{\Sigma}})} and \eqn{\mathcal{N}(\boldsymbol{\mu}, \boldsymbol{\Sigma})} is given by:
\deqn{\dfrac{1}{2}\left(\text{tr}(\boldsymbol{\Sigma}^{-1}\hat{\boldsymbol{\Sigma}}) + \log\left(\dfrac{\det(\boldsymbol{\Sigma})}{\det(\hat{\boldsymbol{\Sigma}})}\right) - p\right),}
where \eqn{\hat{\boldsymbol{\Sigma}}} and \eqn{\boldsymbol{\Sigma}} are the estimated and ground truth covariance matrices, respectively.
}
\references{
Yufeng Zhang, Wanwei Liu, Zhenbang Chen, Ji Wang, and Kenli Li. On the properties of Kullback-Leibler divergence between multivariate gaussian distributions, 2023. \url{https://arxiv.org/abs/2102.05485}
}
