% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{LatentCase}
\alias{LatentCase}
\alias{LatentCase,intData-method}
\title{Latent Case Method for \linkS4class{intData}}
\usage{
LatentCase(Sdt)

\S4method{LatentCase}{intData}(Sdt)
}
\arguments{
\item{Sdt}{An object of class \linkS4class{intData}.}
}
\value{
A character with the latent case.
}
\description{
Latent Case Method for \linkS4class{intData}
}
