% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{LatentDist}
\alias{LatentDist}
\alias{LatentDist,intData-method}
\title{Latent Distribution Method for \linkS4class{intData}}
\usage{
LatentDist(Sdt)

\S4method{LatentDist}{intData}(Sdt)
}
\arguments{
\item{Sdt}{An object of class \linkS4class{intData}.}
}
\value{
A character with the latent distribution(s).
}
\description{
Latent Distribution Method for \linkS4class{intData}
}
