% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{LogRanges}
\alias{LogRanges}
\alias{LogRanges,intData-method}
\title{LogRanges Method for \code{\linkS4class{intData}}}
\usage{
LogRanges(Sdt)

\S4method{LogRanges}{intData}(Sdt)
}
\arguments{
\item{Sdt}{An object of class \code{\linkS4class{intData}}.}
}
\value{
A \code{data.frame} containing the logarithms of the ranges.
}
\description{
LogRanges Method for \code{\linkS4class{intData}}
}
