% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{LowerBounds}
\alias{LowerBounds}
\alias{LowerBounds,intData-method}
\title{Lower Bounds Method for \linkS4class{intData}}
\usage{
LowerBounds(Sdt)

\S4method{LowerBounds}{intData}(Sdt)
}
\arguments{
\item{Sdt}{An object of class \linkS4class{intData}.}
}
\value{
A \code{data.frame} containing the lower bounds of the intervals.
}
\description{
Lower Bounds Method for \linkS4class{intData}
}
