% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_distances.r
\name{Mallows_dist}
\alias{Mallows_dist}
\title{Mallows Distance}
\usage{
Mallows_dist(data, mean_c = NULL, mean_r = NULL)
}
\arguments{
\item{data}{An \linkS4class{intData} object containing the macrodata/interval data}

\item{mean_c}{The mean vector of the centers}

\item{mean_r}{The mean vector of the ranges}
}
\value{
A vector with the squared Mallows distance of each observation.
}
\description{
Calculate the squared Mallows distance between all rows in data and the barycenter.
}
\details{
The squared Mallows distance is defined according to the \code{LatentCase}:
\itemize{
\item \code{"U_id_symmetric"}: The latent variables are identically distributed and symmetric:
\deqn{d_{M}(\boldsymbol{x})^2=(\boldsymbol{c}-\boldsymbol{\mu}_C)^{\top}(\boldsymbol{c}-\boldsymbol{\mu}_C)+\delta(\boldsymbol{r}-\boldsymbol{\mu}_R)^{\top}(\boldsymbol{r}-\boldsymbol{\mu}_R),}
where \eqn{\delta=\mathbb{E}(U^2)/4} is the parameter of the latent variables.
\item \code{"U_id"}: The latent variables are identically distributed:
\deqn{d_{M}(\boldsymbol{x})^2=(\boldsymbol{c}-\boldsymbol{\mu}_C)^{\top}(\boldsymbol{c}-\boldsymbol{\mu}_C)+\delta(\boldsymbol{r}-\boldsymbol{\mu}_R)^{\top}(\boldsymbol{r}-\boldsymbol{\mu}_R)
             +\mathbb{E}(U)(\boldsymbol{c}-\boldsymbol{\mu}_C)^\top(\boldsymbol{r}-\boldsymbol{\mu}_R),}
where \eqn{\delta=\mathbb{E}(U^2)/4} and \eqn{\mathbb{E}(U)} are the parameter of the latent variables.
\item \code{"General"}: The latent variables do not have any nice properties:
\deqn{d_{M}(\boldsymbol{x})^2=(\boldsymbol{c}-\boldsymbol{\mu}_C)^{\top}(\boldsymbol{c}-\boldsymbol{\mu}_C)+(\boldsymbol{r}-\boldsymbol{\mu}_R)^{\top}\boldsymbol{\Delta}(\boldsymbol{r}-\boldsymbol{\mu}_R)
             +(\boldsymbol{c}-\boldsymbol{\mu}_C)^{\top}\boldsymbol{\Psi}(\boldsymbol{r}-\boldsymbol{\mu}_R),}
where:
\itemize{
\item \eqn{\boldsymbol{\Psi}=\text{diag}(\mathbb{E}(U_1),\dots,\mathbb{E}(U_p))},
\item \eqn{\boldsymbol{\Delta}=\text{diag}(\mathbb{E}(U^2_1),\dots,\mathbb{E}(U^2_p))/4}.
}
}
}
\examples{
data(creditcard)
credit_card_int <- creditcard$intData

credit_card_dist<-Mallows_dist(credit_card_int)
}
\references{
Oliveira, M. R., Pinheiro, D., & Oliveira, L. (2025).
Location and association measures for interval-valued data based on Mallows' distance.
arXiv preprint arXiv:2407.05105. \url{https://arxiv.org/abs/2407.05105}
}
