% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_utils.r
\name{meanU}
\alias{meanU}
\title{Compute Mean Latent Variables}
\usage{
meanU(
  LatentDist = c("Unif", "Triang", "TNorm", "InvTri", "Beta", "KDE", "Degenerated"),
  TriangParam = 0,
  BetaParam.a = 1,
  BetaParam.b = 1,
  Umicro = NULL,
  p = NULL
)
}
\arguments{
\item{LatentDist}{A string or vector of strings specifying the distribution(s) of the latent variables. If the variables are identically distributed it can be one of (\code{"Unif"},\code{"Triang"},\code{"TNorm"},\code{"InvTri"},\code{"Beta"},\code{"KDE"},\code{"Degenerated"}), if not a vector must be provided with the distribution for each variable.}

\item{TriangParam}{Mode of the triangular distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{0}.}

\item{BetaParam.a}{Parameter alpha of the Beta distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{1}.}

\item{BetaParam.b}{Parameter beta of the Beta distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{1}.}

\item{Umicro}{Latent microdata observations. Needed if \code{LatentDist="KDE"}.}

\item{p}{Number of variables.}
}
\value{
Either a diagonal matrix with the mean of each variable or a value if the variables are identically distributed.
}
\description{
Obtain the mean of the latent variables inherent to the macrodata.
}
\keyword{internal}
