% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro2intData.r
\name{micro2intData}
\alias{micro2intData}
\title{Aggregate Microdata into Interval-Valued Data}
\usage{
micro2intData(
  microdata,
  agrby,
  agrcrt = "minmax",
  LatentParam = NULL,
  LatentCase = c("U_id_symmetric", "U_id", "General"),
  LatentDist = c("Unif", "Triang", "TNorm", "InvTri", "Beta", "KDE", "Degenerated"),
  TriangParam = 0,
  BetaParam.a = 1,
  BetaParam.b = 1,
  estimate.DistParam = FALSE,
  removeDegenerate = FALSE
)
}
\arguments{
\item{microdata}{A data frame containing the microdata. All columns should be numeric.}

\item{agrby}{A factor used to specify the grouping of the microdata for aggregation.}

\item{agrcrt}{A string or numeric vector of length 2 specifying the aggregation criterion. The default is \code{"minmax"}, which takes the minimum and maximum values for each variable. If a numeric vector is provided, it should specify the lower and upper percentiles for aggregation (e.g., \code{c(0.05, 0.95)}).}

\item{LatentParam}{(Optional) A list with the parameters of the latent variables.
Expects a list with a single number if \code{LatentCase} is \code{"U_id_symmetric"}, a list of two numbers if \code{LatentCase} is \code{"U_id"}, and a list of two matrices if \code{LatentCase} is \code{"General"}.}

\item{LatentCase}{A string specifying which of the three scenarios applies to the latent variables:
\itemize{
\item \code{"U_id_symmetric"}: The case where the latent variables are identically distributed and symmetric.
\item \code{"U_id"}: The case where the latent variables are identically distributed.
\item \code{"General"}: The case where the latent variables do not have any nice properties.
}
Defaults to \code{"U_id_symmetric"}.}

\item{LatentDist}{A string or vector of strings specifying the distribution(s) of the latent variables. If the variables are identically distributed it can be one of (\code{"Unif"}, \code{"Triang"}, \code{"TNorm"}, \code{"InvTri"}, \code{"Beta"}, \code{"KDE"}, \code{"Degenerated"}), if not a vector must be provided with the distribution for each variable.
The default is \code{"Unif"} if \code{LatentCase="U_id_symmetric"}, and \code{"KDE"} if \code{LatentCase="General"}.}

\item{TriangParam}{Mode of the triangular distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{0}.}

\item{BetaParam.a}{Parameter alpha of the Beta distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{1}.}

\item{BetaParam.b}{Parameter beta of the Beta distribution. If the latent variables are identically distributed, it is only necessary to provide a number, if not a vector is needed.
The default is \code{1}.}

\item{estimate.DistParam}{Logical parameter indicating if estimation of the parameters of the latent distributions should be performed. Can only be set to TRUE if \code{LatentCase="General"}.
The default is \code{FALSE}.}

\item{removeDegenerate}{Logical parameter indicating if observations with at least one degenerate interval should be removed. The default is \code{FALSE}.}
}
\value{
An \code{\linkS4class{intData}} object containing the aggregated interval-valued data. If \code{removeDegenerate} is \code{TRUE}, observations with degenerate intervals are removed, and \code{NULL} is returned if all observations are removed.
}
\description{
Aggregates microdata from a data frame into interval-valued data using various criteria and latent distribution settings.
}
\details{
This function processes a data frame of microdata and aggregates it into interval-valued data according to the specified grouping factor and aggregation criteria.
It can handle different latent distribution cases and parameter settings.

If some rows contain invalid (non-finite or missing) values, those rows are removed before aggregation. Degenerate intervals (i.e., intervals whose lower and upper bounds are equal) can be removed by setting \code{removeDegenerate} to \code{TRUE}.
}
\examples{
data(creditcard)
CreditCard_microdata <- creditcard$microdata

# Define grouping variable for microdata aggregation
credit_agrby <- factor(paste(CreditCard_microdata$Name, CreditCard_microdata$Month, sep = "_"))

# Create intData object by aggregating microdata using the default minmax criterion 
# and using KDE for estimation of the latent distribution in the general case
credit_agr <- micro2intData(CreditCard_microdata[,3:7],
                            agrby = credit_agrby,
                            LatentCase = "General")

}
\references{
Adapted from package \code{MAINT.Data} (\url{https://cran.r-project.org/package=MAINT.Data}).
}
