% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{NbMicroUnits}
\alias{NbMicroUnits}
\alias{NbMicroUnits,intData-method}
\title{Number of Micro Units Method for \linkS4class{intData}}
\usage{
NbMicroUnits(x)

\S4method{NbMicroUnits}{intData}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{intData}.}
}
\value{
An integer specifying the number of micro units.
}
\description{
Number of Micro Units Method for \linkS4class{intData}
}
