% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{NMicro}
\alias{NMicro}
\alias{NMicro,intData-method}
\title{Number of Micro Units Method for \code{\linkS4class{intData}}}
\usage{
NMicro(x)

\S4method{NMicro}{intData}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{intData}}.}
}
\value{
An integer specifying the number of micro units.
}
\description{
Number of Micro Units Method for \code{\linkS4class{intData}}
}
