% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{plot,intData,intData-method}
\alias{plot,intData,intData-method}
\alias{plot,intData,missing-method}
\title{Plot Method for Two \linkS4class{intData} Objects}
\usage{
\S4method{plot}{intData,intData}(
  x,
  y,
  type = c("crosses", "rectangles", "crosses2"),
  append = FALSE,
  palette = rainbow(x@NObs),
  ...
)

\S4method{plot}{intData,missing}(
  x,
  casen = NULL,
  layout = c("vertical", "horizontal"),
  append = FALSE,
  ...
)
}
\arguments{
\item{x}{An \linkS4class{intData} object.}

\item{y}{An \linkS4class{intData} object to plot on the y-axis.}

\item{type}{The type of plot to generate: "crosses" or "rectangles" or "crosses2". Default is "crosses".}

\item{append}{Logical, if \code{TRUE}, the plot is added to the current plot.}

\item{palette}{A vector with colors for each observation.}

\item{...}{Additional graphical parameters.}

\item{casen}{A vector specifying the case numbers to plot. Default is \code{NULL}.}

\item{layout}{The layout of the plot: "vertical" or "horizontal".}
}
\value{
A plot showing the relationship between the two \linkS4class{intData} objects.

A plot showing the intervals of the \linkS4class{intData} object.
}
\description{
Plots one \linkS4class{intData} object against another, with options to visualize the intervals as crosses or rectangles.

Plots a single \linkS4class{intData} object, either in a vertical or horizontal layout.
}
