% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_Shapley.r
\name{plot_int_Shapley_inter}
\alias{plot_int_Shapley_inter}
\title{Plot Shapley interaction indices}
\usage{
plot_int_Shapley_inter(
  x,
  abbrev = 10,
  title = NULL,
  legend = TRUE,
  text_size = 22
)
}
\arguments{
\item{x}{A \eqn{p \times p} matrix containing the Shapley interaction indices of a single observation.}

\item{abbrev}{Integer. If \code{abbrev.var} \eqn{> 0}, variable names are abbreviated using abbreviate with \code{minlenght = abrev}.}

\item{title}{Character. Title of the plot.}

\item{legend}{Logical. If TRUE (default), a legend is plotted.}

\item{text_size}{Integer. Size of the text in the plot}
}
\value{
Returns a figure consisting of two panels. The right panel shows the Shapley values, and the left panel the Shapley interaction indices.
}
\description{
Plot Shapley interaction indices
}
\examples{
# Create intData object
data(creditcard)
credit_card_int <- creditcard$intData

# Estimate the mean and covariance matrix
credit_card_IMCD <- IMCD(credit_card_int, 
                         m = floor(nrow(credit_card_int)*0.75), 
                         cutoff = "farness", 
                         cutoff_lvl = 0.9)

# Compute Shapley interaction indices
credit_card_shap_inter <- int_Shapley_interaction(credit_card_int, 
                                                  mean_c = credit_card_IMCD$mean_IMCD_c, 
                                                  mean_r = credit_card_IMCD$mean_IMCD_r, 
                                                  cov = credit_card_IMCD$cov_IMCD)

# Plot Shapley interaction for 1st observation
plot_int_Shapley_inter(credit_card_shap_inter[[1]])
}
\references{
Adapted from package \code{ShapleyOutlier} (\url{https://CRAN.R-project.org/package=ShapleyOutlier}).
}
