% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_Shapley.r
\name{plot_radar_int_Shapley}
\alias{plot_radar_int_Shapley}
\title{Radar plot of Shapley values for interval-valued data.}
\usage{
plot_radar_int_Shapley(shapley, palette = NULL, sort.obs = FALSE)
}
\arguments{
\item{shapley}{A \eqn{n \times p} matrix containing the Shapley values of \eqn{n} observations and \eqn{p} variables.}

\item{palette}{A vector of palette for each observation. Default is black.}

\item{sort.obs}{Logical. If \code{TRUE} (default), observations are sorted according to their squared (robust) Interval-Mahalanobis distance.}
}
\value{
Returns a radar plot that displays the Shapley values (\code{\link{int_Shapley}}) for each observation.
}
\description{
Radar plot of Shapley values for interval-valued data.
}
\examples{
# Create intData object
data(creditcard)
credit_card_int <- creditcard$intData

# Estimate the mean and covariance matrix
credit_card_IMCD <- IMCD(credit_card_int, 
                         m = floor(nrow(credit_card_int)*0.75), 
                         cutoff = "farness", 
                         cutoff_lvl = 0.9)

# Detect outliers using farness cutoff
credit_card_outliers <- int_outliers(credit_card_IMCD$robust_dist, 
                                     cutoff = "farness", 
                                     cutoff_lvl = 0.9)

# Compute Shapley values
credit_card_shapley <- int_Shapley(credit_card_int, 
                                   mean_c = credit_card_IMCD$mean_IMCD_c,
                                   mean_r = credit_card_IMCD$mean_IMCD_r, 
                                   cov = credit_card_IMCD$cov_IMCD)

# colors
outliers_colors <- rep('black',credit_card_int@NObs)
names(outliers_colors) <- rownames(credit_card_int)
outliers_colors[credit_card_outliers$outliers_names] = '#009de0'

plot_radar_int_Shapley(credit_card_shapley, palette = outliers_colors)
}
