% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_Shapley.r
\name{plot_tile_int_Shapley}
\alias{plot_tile_int_Shapley}
\title{Tileplot of Shapley values for interval-valued data.}
\usage{
plot_tile_int_Shapley(
  shapley,
  outliers = NULL,
  rotate_x = TRUE,
  abbrev.var = FALSE,
  abbrev.obs = FALSE,
  sort.var = FALSE,
  sort.obs = FALSE,
  show_values = FALSE
)
}
\arguments{
\item{shapley}{A \eqn{n \times p} matrix containing the Shapley values of \eqn{n} observations and \eqn{p} variables.}

\item{outliers}{A list containing the outliers' names as returned by \code{\link{int_outliers}}. If \code{outliers} is not \code{NULL} (default), only the outliers are highlighted in the plot.}

\item{rotate_x}{Logical. If \code{TRUE} (default), the x-axis labels are rotated.}

\item{abbrev.var}{Integer. If \code{abbrev.var} \eqn{> 0}, column names are abbreviated using abbreviate with \code{minlenght = abrev.var}.}

\item{abbrev.obs}{Integer. If \code{abbrev.obs} \eqn{> 0}, row names are abbreviated using abbreviate with \code{minlenght = abrev.obs}.}

\item{sort.var}{Logical. If \code{TRUE}, variables are sorted according to the distance.}

\item{sort.obs}{Logical. If \code{TRUE}, observations are sorted according to their squared Interval-Mahalanobis distance.}

\item{show_values}{Logical. If \code{TRUE}, the Shapley values are displayed in each tile.}
}
\value{
Returns a tileplot that displays the Shapley values (\code{\link{int_Shapley}}) for each observation and variable. Optionally, only the outliers are highlighted in the plot.
}
\description{
Tileplot of Shapley values for interval-valued data.
}
\examples{
# Create intData object
data(creditcard)
credit_card_int <- creditcard$intData

# Estimate the mean and covariance matrix
credit_card_IMCD <- IMCD(credit_card_int, 
                         m = floor(nrow(credit_card_int)*0.75), 
                         cutoff = "farness", 
                         cutoff_lvl = 0.9)

# Detect outliers using farness cutoff
credit_card_outliers <- int_outliers(credit_card_IMCD$robust_dist, 
                                     cutoff = "farness", 
                                     cutoff_lvl = 0.9)

# Compute Shapley values
credit_card_shapley <- int_Shapley(credit_card_int, 
                                   mean_c = credit_card_IMCD$mean_IMCD_c, 
                                   mean_r = credit_card_IMCD$mean_IMCD_r, 
                                   cov = credit_card_IMCD$cov_IMCD)

plot_tile_int_Shapley(credit_card_shapley, 
                     outliers = credit_card_outliers, 
                     sort.var = TRUE, 
                     sort.obs = TRUE)
}
\references{
Adapted from package \code{ShapleyOutlier} (\url{https://CRAN.R-project.org/package=ShapleyOutlier}).
}
