% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{print,summaryintData-method}
\alias{print,summaryintData-method}
\title{Print Method for Summary \linkS4class{intData}}
\usage{
\S4method{print}{summaryintData}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summaryintData}.}

\item{...}{Additional arguments passed to print.}
}
\value{
The object itself, returned invisibly. Called for its side effects (printing).
}
\description{
Print Method for Summary \linkS4class{intData}
}
