% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{Ranges}
\alias{Ranges}
\alias{Ranges,intData-method}
\title{Ranges Method for \linkS4class{intData}}
\usage{
Ranges(Sdt)

\S4method{Ranges}{intData}(Sdt)
}
\arguments{
\item{Sdt}{An object of class \linkS4class{intData}.}
}
\value{
A \code{data.frame} containing the ranges of the intervals.
}
\description{
Ranges Method for \linkS4class{intData}
}
