% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{row.names,intData-method}
\alias{row.names,intData-method}
\title{Row.Names Method for \linkS4class{intData}}
\usage{
\S4method{row.names}{intData}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{intData}.}
}
\value{
A character vector of row names.
}
\description{
Row.Names Method for \linkS4class{intData}
}
