% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{show,intData-method}
\alias{show,intData-method}
\alias{show,summaryintData-method}
\title{Show Method for \linkS4class{intData}}
\usage{
\S4method{show}{intData}(object)

\S4method{show}{summaryintData}(object)
}
\arguments{
\item{object}{An object of class \code{summaryintData}.}
}
\value{
The object itself, returned invisibly. Called for its side effects (printing).
}
\description{
Show Method for \linkS4class{intData}

Show Method for Summary \linkS4class{intData}
}
