% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\docType{class}
\name{summaryintData-class}
\alias{summaryintData-class}
\title{Summary Interval Data Class}
\description{
A class to represent the summary of interval data.
}
\section{Slots}{

\describe{
\item{\code{Centersumar}}{A table summarizing the centers.}

\item{\code{Rngsumar}}{A table summarizing the ranges.}
}}

