% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{UpperBounds}
\alias{UpperBounds}
\alias{UpperBounds,intData-method}
\title{Upper Bounds Method for \linkS4class{intData}}
\usage{
UpperBounds(Sdt)

\S4method{UpperBounds}{intData}(Sdt)
}
\arguments{
\item{Sdt}{An object of class \linkS4class{intData}.}
}
\value{
A \code{data.frame} containing the upper bounds of the intervals.
}
\description{
Upper Bounds Method for \linkS4class{intData}
}
