% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCD.r
\name{bigIMCD}
\alias{bigIMCD}
\title{Obtain unweighted estimates for data with > 600 observations}
\usage{
bigIMCD(m, p, n, data)
}
\arguments{
\item{m}{An integer specifying number of observations to use}

\item{p}{An integer specifying the number of columns in X}

\item{n}{An integer specifying the number of total observations}

\item{data}{An \linkS4class{intData} object containing the macrodata/interval data}
}
\value{
A list of estimated location and scatter
}
\description{
Obtain unweighted estimates for data with > 600 observations
}
\keyword{internal}
