% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCD.r
\name{c_step}
\alias{c_step}
\title{Perform single iteration of C-step}
\usage{
c_step(z, m, data)
}
\arguments{
\item{z}{A vector of 0 and 1, indicating which observations should be considered for the calculation}

\item{m}{An integer specifying number of observations to use}

\item{data}{An \linkS4class{intData} object containing the macrodata/interval data}
}
\value{
A list of z, covariance, barycenter and robust distances
}
\description{
Perform single iteration of C-step
}
\keyword{internal}
