% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{entrecampos_air_quality}
\alias{entrecampos_air_quality}
\title{Entrecampos Air Quality Dataset}
\format{
A list with the following components:
\describe{
\item{\code{microdata_raw}}{A data frame with \code{26304} rows and \code{11} columns. It contains the raw microdata, with individual measurements of each variable for all observations.}
\item{\code{microdata_transformed}}{A data frame with \code{26304} rows and \code{10} columns. It contains the microdata, with individual measurements of each variable for all observations. Logarithmic transformations were applied to all variables and interpolation to deal with missing values.}
\item{\code{min_max}}{A data frame with \code{1096} rows and \code{17} columns. Each row corresponds to a different observation, and each column gives the minimum and maximum values for each variable. The first column corresponds to the day, the next 8 to the minimum and the last 8 to the maximum.}
\item{\code{intData}}{An \linkS4class{intData} object, constructed using KDE for estimating the parameters of the latent distributions.}
}
}
\usage{
data(entrecampos_air_quality)
}
\description{
This dataset contains interval data of air pollutants' concentrations, including min-max values and microdata.
This air quality dataset was obtained from a monitoring station in Entrecampos, Lisbon.
It is composed of 9 pollutants' concentration measures in µg/m3 during the years 2019, 2020, and 2021: sulphur dioxide (SO2), particles < 10µm, ozone (O3), nitrogen dioxide (NO2), carbon monoxide (CO), benzene (C6H6), particles < 2.5µm, nitrogen oxides (NOx), and nitrogen monoxide (NO).
For the \code{microdata_transformed}, \code{min_max}, and \code{intData}, the pollutant "benzene" was removed due to a high number of missing values.
The aggregation of the microdata was done by day.
}
\examples{
data(entrecampos_air_quality)
head(entrecampos_air_quality$microdata_raw)
head(entrecampos_air_quality$microdata_transformed)
head(entrecampos_air_quality$min_max)
head(entrecampos_air_quality$intData)

}
\references{
This data was retrieved from the Portuguese Environment Agency database available at \url{https://qualar.apambiente.pt/}.
}
\keyword{datasets}
