% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_distances.r
\name{IMah_dist_pairs}
\alias{IMah_dist_pairs}
\title{Interval-Mahalanobis distance for all pairs}
\usage{
IMah_dist_pairs(data, cov = NULL)
}
\arguments{
\item{data}{An \linkS4class{intData} object containing the macrodata/interval data}

\item{cov}{The symbolic covariance matrix}
}
\value{
A matrix with the squared Interval-Mahalanobis distance of each pair of observations.
}
\description{
Calculate the squared Interval-Mahalanobis distance of all pairs of observations in the data.
}
\details{
The squared Interval-Mahalanobis distance is defined according to the \code{LatentCase}:
\itemize{
\item \code{"U_id_symmetric"}: The latent variables are identically distributed and symmetric:
\deqn{d_{IMah}(\boldsymbol{x}_i,\boldsymbol{x}_j)^2=(\boldsymbol{c}_i-\boldsymbol{c}_j)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{c}_i-\boldsymbol{c}_j)+\delta(\boldsymbol{r}_i-\boldsymbol{r}_j)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{r}_i-\boldsymbol{r}_j),}
where \eqn{\delta=\mathbb{E}(U^2)/4} is the parameter of the latent variables.
\item \code{"U_id"}: The latent variables are identically distributed:
\deqn{d_{IMah}(\boldsymbol{x}_i,\boldsymbol{x}_j)^2=(\boldsymbol{c}_i-\boldsymbol{c}_j)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{c}_i-\boldsymbol{c}_j)+\delta(\boldsymbol{r}_i-\boldsymbol{r}_j)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{r}_i-\boldsymbol{r}_j)\\
             +\dfrac{\mathbb{E}(U)}{2}(\boldsymbol{c}_i-\boldsymbol{c}_j)^\top\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{r}_i-\boldsymbol{r}_j)+\dfrac{\mathbb{E}(U)}{2}(\boldsymbol{r}_i-\boldsymbol{r}_j)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{c}_i-\boldsymbol{c}_j),}
where \eqn{\delta=\mathbb{E}(U^2)/4} and \eqn{\mathbb{E}(U)} are the parameter of the latent variables.
\item \code{"General"}: The latent variables do not have any nice properties:
\deqn{d_{IMah}(\boldsymbol{x}_i,\boldsymbol{x}_j)^2=(\boldsymbol{c}_i-\boldsymbol{c}_j)^{\top}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{c}_i-\boldsymbol{c}_j)+\dfrac{1}{4}(\boldsymbol{r}_i-\boldsymbol{r}_j)^{\top}\left(\boldsymbol{\mathfrak{E}}_{UU}\bullet\boldsymbol{\Sigma}_{B}^{-1}\right)(\boldsymbol{r}_i-\boldsymbol{r}_j)\\
             +\dfrac{1}{2}(\boldsymbol{c}_i-\boldsymbol{c}_j)^{\top}\boldsymbol{\Sigma}_{B}^{-1}\boldsymbol{\Psi}(\boldsymbol{r}_i-\boldsymbol{r}_j)+\dfrac{1}{2}(\boldsymbol{r}_i-\boldsymbol{r}_j)^{\top}\boldsymbol{\Psi}\boldsymbol{\Sigma}_{B}^{-1}(\boldsymbol{c}_i-\boldsymbol{c}_j),}
where:
\itemize{
\item \eqn{\boldsymbol{\Psi}=\text{diag}(\mathbb{E}(U_1),\dots,\mathbb{E}(U_p))},
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ij}=\mathcal{E}(U_i,U_j)}, \eqn{i\neq j}, with \eqn{\mathcal{E}(U_i,U_j)=\int_0^1 F_{U_i}^{-1}(t) F_{U_j}^{-1}(t) \, dt},
\item \eqn{[\boldsymbol{\mathfrak{E}}_{UU}]_{ii}=\mathbb{E}(U_i^2)}, \eqn{i,j=1,\dots,p},
\item \eqn{\bullet} denotes the Schur (or entrywise) product of matrices.
}
}
}
\examples{
data(creditcard)
credit_card_int <- creditcard$intData

credit_card_dist<-IMah_dist_pairs(credit_card_int)
}
\references{
Loureiro, C. P., Oliveira, M. R., Brito, P., & Oliveira, L. (2026).
Minimum Covariance Determinant Estimator and Outlier Detection for Interval-valued Data.
arXiv preprint arXiv:2604.26769. \url{https://arxiv.org/abs/2604.26769}
}
