% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCD.r
\name{IMCD}
\alias{IMCD}
\title{Interval-valued data Minimum Covariance Determinant (IMCD) estimation}
\usage{
IMCD(
  data,
  m = 0,
  cutoff = c("farness", "adjbox", "chi-squared", "F-dist", "raw"),
  cutoff_lvl = NULL
)
}
\arguments{
\item{data}{An \linkS4class{intData} object containing the interval-valued dataset (macrodata).}

\item{m}{An integer specifying the subset size to use for the estimation. Defaults to \code{floor(0.75*n)}.}

\item{cutoff}{Indicates which cutoff should be considered for reweighting the estimates:
\itemize{
\item \code{"chi-squared"}: The traditional 97.5\\% Chi-Squared quantile.
\item \code{"raw"}: No reweighting.
\item \code{"adjbox"}: Adjusted Boxplots (package \code{robustbase}).
\item \code{"F-dist"}: The quantile of the scaled F distribution (adapted from package \code{CerioliOutlierDetection}).
\item \code{"farness"}: "Farness" is estimated from the robust distance (adapted from package \code{cellWise}).
}
Defaults to \code{"farness"}.}

\item{cutoff_lvl}{A numeric value specifying the level of the cutoff to be used.
\itemize{
\item If \code{cutoff="chi-squared"}, \code{cutoff_lvl} is the quantile of the Chi-squared distribution (default is 0.975).
\item If \code{cutoff="adjbox"}, \code{cutoff_lvl} is the coefficient for the adjusted boxplot (default is 1.5).
\item If \code{cutoff="F-dist"}, \code{cutoff_lvl} is the quantile of the F-distribution (default is 0.975).
\item If \code{cutoff="farness"}, \code{cutoff_lvl} represents the threshold for farness, with a default of 0.99.
\item If \code{cutoff="raw"}, \code{cutoff_lvl} is ignored.
}
If no value is provided, the function uses the default values associated with each cutoff method.}
}
\value{
A list containing the robustly estimated parameters:
\item{\code{mean_IMCD_c}}{Estimated mean of the centers of the interval data.}
\item{\code{mean_IMCD_r}}{Estimated mean of the ranges of the interval data.}
\item{\code{cov_IMCD}}{Estimated covariance (scatter) matrix (\code{\link{int_cov}}) for the data.}
\item{\code{final_z}}{Binary vector indicating the inclusion of each observation in the reweighted subset.}
\item{\code{cutoff}}{The cutoff method used for reweighting.}
\item{\code{cutoff_value}}{Cutoff value used for reweighting.}
\item{\code{robust_dist}}{Robust distances (\code{\link{IMah_dist}}) for each observation.}
\item{\code{farness_probs}}{Farness probabilities (if \code{cutoff} is set to \code{"farness"}).}
}
\description{
Applies an adaptation of the FAST-MCD algorithm to estimate location and scatter for interval-valued data.
}
\examples{
# Example using creditcard dataset
data(creditcard)
credit_card_int <- creditcard$intData

credit_card_IMCD <- IMCD(credit_card_int, floor(0.75*credit_card_int@NObs), "farness", 0.9)
}
\references{
Loureiro, C. P., Oliveira, M. R., Brito, P., & Oliveira, L. (2026).
Minimum Covariance Determinant Estimator and Outlier Detection for Interval-valued Data.
arXiv preprint arXiv:2604.26769. \url{https://arxiv.org/abs/2604.26769}

Adapted from \url{https://github.com/frankp-0/fastMCD}.

The case \code{cutoff=="F-dist"} is adapted from package \code{CerioliOutlierDetection} (\url{https://cran.r-project.org/package=CerioliOutlierDetection}).
}
