% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\docType{class}
\name{intData-class}
\alias{intData-class}
\title{Interval Data Class}
\description{
A class to represent interval data.
}
\section{Slots}{

\describe{
\item{\code{Centers}}{A data frame of centers of the intervals.}

\item{\code{Ranges}}{A data frame of ranges of the intervals.}

\item{\code{LatentParam}}{A list with the parameters of the latent variables.}

\item{\code{LatentCase}}{A string specifying which of the three scenarios applies to the latent variables:
\itemize{
\item \code{"General"}: The case where the latent variables do not have any nice properties.
\item \code{"U_id"}: The case where the latent variables are identically distributed.
\item \code{"U_id_symmetric"}: The case where the latent variables are identically distributed and symmetric.
}
Defaults to \code{"U_id_symmetric"}.}

\item{\code{LatentDist}}{A string or vector of strings specifying the distribution(s) of the latent variables. If the variables are identically distributed it can be one of (\code{"Unif"},\code{"Triang"},\code{"TNorm"},\code{"InvTri"},\code{"Beta"},\code{"KDE"},\code{"Degenerated"}), if not, it is a vector with the distribution for each variable.}

\item{\code{ObsNames}}{A character vector of observation names.}

\item{\code{VarNames}}{A character vector of variable names.}

\item{\code{NObs}}{A numeric value indicating the number of observations.}

\item{\code{NIVar}}{A numeric value indicating the number of interval variables.}

\item{\code{NbMicroUnits}}{An integer indicating the number of micro units.}
}}

\references{
Oliveira, M. R., Pinheiro, D., & Oliveira, L. (2025).
Location and association measures for interval-valued data based on Mallows' distance.
arXiv preprint arXiv:2407.05105. \url{https://arxiv.org/abs/2407.05105}

Adapted from package \code{MAINT.Data} (\url{https://cran.r-project.org/package=MAINT.Data}).
}
