% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{LatentParam}
\alias{LatentParam}
\alias{LatentParam,intData-method}
\title{Latent Parameters Method for \linkS4class{intData}}
\usage{
LatentParam(Sdt)

\S4method{LatentParam}{intData}(Sdt)
}
\arguments{
\item{Sdt}{An object of class \linkS4class{intData}.}
}
\value{
A list with the latent parameters.
}
\description{
Latent Parameters Method for \linkS4class{intData}
}
