% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCD.r
\name{pick10}
\alias{pick10}
\title{Choose the 10 best estimates after iterating twice through initial sets}
\usage{
pick10(z_all, m, data)
}
\arguments{
\item{z_all}{A 2D matrix where each row specifies a subset of observations}

\item{m}{An integer specifying number of observations to use}

\item{data}{An \linkS4class{intData} object containing the macrodata/interval data}
}
\value{
A list of z, covariance, barycenter and robust distances
}
\description{
Choose the 10 best estimates after iterating twice through initial sets
}
\keyword{internal}
