% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_Shapley.r
\name{plot_bar_int_Shapley_decomp}
\alias{plot_bar_int_Shapley_decomp}
\title{Barplot of Shapley value decomposition into contributions of (Centers, Ranges, and CrossCentersRanges) for interval-valued data.}
\usage{
plot_bar_int_Shapley_decomp(
  shapley_decomp,
  palette = NULL,
  rotate_x = TRUE,
  abbrev.obs = 20,
  sort.obs = TRUE,
  plot_IMah = FALSE
)
}
\arguments{
\item{shapley_decomp}{A list of matrices containing the Shapley value decomposition into contributions of (Centers, Ranges, and CrossCentersRanges) for each observation.}

\item{palette}{A vector with colors for each feature. If \code{palette} is \code{NULL} (default), the colors are generated using \code{RColorBrewer}.}

\item{rotate_x}{Logical. If \code{TRUE} (default), the x-axis labels are rotated.}

\item{abbrev.obs}{Integer. If \code{abbrev.obs} \eqn{> 0}, row names are abbreviated using abbreviate with \code{minlenght = abbrev.obs}.}

\item{sort.obs}{Logical. If \code{TRUE} (default), observations are sorted according to their total Shapley value.}

\item{plot_IMah}{Logical. If \code{TRUE}, the Interval-Mahalanobis distance (sum of all Shapley values) will be included in the plot.}
}
\value{
Returns a barplot that displays the Shapley value decomposition into contributions of (Centers, Ranges, and CrossCentersRanges) for each observation.
}
\description{
Barplot of Shapley value decomposition into contributions of (Centers, Ranges, and CrossCentersRanges) for interval-valued data.
}
\examples{
# Create intData object
data(creditcard)
credit_card_int <- creditcard$intData

# Compute Shapley decomposition into contributions of Centers, Ranges, and CrossCentersRanges
# based on IMCD estimates of mean and covariance matrix
credit_card_shap_decomp <- int_Shapley_decomp(credit_card_int)

# Plot Shapley decomposition with contributions of Centers, Ranges, and CrossCentersRanges
plot_bar_int_Shapley_decomp(credit_card_shap_decomp, palette = rainbow(credit_card_int@NIVar))
}
