% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_dist_dist}
\alias{plot_dist_dist}
\title{Distance-Distance plot for interval-valued data.}
\usage{
plot_dist_dist(
  class_dist,
  class_cutoff = NULL,
  class_cutoff_label = NULL,
  rob_dist,
  rob_cutoff = NULL,
  rob_cutoff_label = NULL,
  obs_names = NULL,
  ggplotly = TRUE,
  color_class = NULL,
  color_label = NULL,
  palette = NULL,
  shape_class = NULL,
  shape_label = NULL,
  label_obs = NULL
)
}
\arguments{
\item{class_dist}{A numeric vector containing the classical distances for each observation.}

\item{class_cutoff}{Numeric. The cutoff value for the classical distances.}

\item{class_cutoff_label}{Character. Label for the classical cutoff. If NULL (default), no legend for the classical cutoff is shown.}

\item{rob_dist}{A numeric vector containing the robust distances for each observation.}

\item{rob_cutoff}{Numeric. The cutoff value for the robust distances.}

\item{rob_cutoff_label}{Character. Label for the robust cutoff. If NULL (default), no legend for the robust cutoff is shown.}

\item{obs_names}{A character vector containing the names of the observations. If NULL (default), the names are taken from the names of class_dist.}

\item{ggplotly}{Logical. If \code{TRUE} (default), the plot is converted to an interactive \link[plotly:plotly]{plotly::plotly} object.}

\item{color_class}{A vector indicating the color class of each observation. If NULL (default), all points have the same color.}

\item{color_label}{Character. Label for the color class. If NULL (default), no legend for the color class is shown.}

\item{palette}{A vector with colors for each color class. If NULL (default), default \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} colors are used.}

\item{shape_class}{A vector indicating the shape class of each observation. If NULL (default), all points have the same shape.}

\item{shape_label}{Character. Label for the shape class. If NULL (default), no legend for the shape class is shown.}

\item{label_obs}{A vector with the names of the observations to be labeled in the plot when \code{ggplotly = FALSE}. Default is NULL.}
}
\value{
Returns a Distance-Distance plot that displays the classical distances against the robust distances for each observation, highlighting outliers.
}
\description{
Distance-Distance plot for interval-valued data.
}
\examples{
#Create intData object
data(creditcard)
credit_card_int <- creditcard$intData

#Estimate the mean and covariance matrix
credit_card_IMCD<-IMCD(credit_card_int, floor(nrow(credit_card_int)*0.75), "farness", 0.9)
credit_card_outliers <- int_outliers(credit_card_IMCD$robust_dist, 
                                           p=credit_card_int@NIVar, cutoff_lvl = 0.9)

#Plot Distance-Distance plot
class_dist <- IMah_dist(credit_card_int, z=rep(1,credit_card_int@NObs))
class_outliers <- int_outliers(class_dist,cutoff = "adjbox",p=p,cutoff_lvl = 1.5)
credit_card_is_outliers <- as.character(credit_card_outliers$is_outlier)
credit_card_is_outliers[credit_card_outliers$is_outlier] <- "Outlier"
credit_card_is_outliers[!credit_card_outliers$is_outlier] <- "Inlier"
plot_dist_dist(class_dist, class_outliers$cutoff_value[2], "1.5 adjusted boxplot",
              credit_card_IMCD$robust_dist, credit_card_outliers$cutoff_value, "0.9 farness",
              color_class = credit_card_is_outliers, palette = c("grey50", "red"))
}
