% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intData_class.r
\name{[,intData-method}
\alias{[,intData-method}
\alias{[,intData,ANY,ANY-method}
\title{Subset an \linkS4class{intData} Object}
\usage{
\S4method{[}{intData}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An \linkS4class{intData} object.}

\item{i}{Row indices or names to subset. Defaults to all rows.}

\item{j}{Column indices or names to subset. Defaults to all columns.}

\item{...}{Additional arguments (not used).}

\item{drop}{Logical, passed to the underlying \code{[}. Defaults to \code{TRUE}.}
}
\value{
An \linkS4class{intData} object containing the specified subset of rows and columns.
}
\description{
Extract a subset of rows and columns from an \linkS4class{intData} object.
}
