% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{spotify_tracks}
\alias{spotify_tracks}
\title{Spotify Tracks Dataset}
\format{
A list with the following components:
\describe{
\item{\code{microdata}}{A data frame with \code{81033} rows and \code{20} columns. It contains the microdata, with individual measurements of each variable for all observations.}
\item{\code{microdata_transformed}}{A data frame with \code{81033} rows and \code{20} columns. It contains the transformed microdata, with individual measurements of each variable for all observations. Logarithmic transformations were applied to "loudness" and "tempo". "duration_ms" in milliseconds was converted to "duration" in minutes. "popularity" was scaled to the range \verb{[0,1]}.}
\item{\code{intData_minmax}}{An \linkS4class{intData} object with \code{111} interval-valued observations and \code{11} variables, constructed using min-max aggregation based on the transformed microdata.}
\item{\code{intData_trimmed}}{An \linkS4class{intData} object with \code{111} interval-valued observations and \code{11} variables, constructed using trimmed aggregation (\verb{1\\\%} trimming) based on the transformed microdata.}
}
}
\usage{
data(spotify_tracks)
}
\description{
This dataset contains interval data of Spotify tracks' audio features, including min-max values and trimmed intervals, as well as the microdata.
It is composed of 11 audio features: duration, danceability, energy, loudness, speechiness, acousticness, instrumentalness, liveness, valence, tempo, and popularity.
The aggregation of the microdata was done by track genre.
}
\examples{
data(spotify_tracks)
head(spotify_tracks$intData_minmax)
head(spotify_tracks$intData_trimmed)
head(spotify_tracks$microdata)
head(spotify_tracks$microdata_transformed)

}
\references{
This data was retrieved from Kaggle, available at \url{https://www.kaggle.com/datasets/maharshipandya/-spotify-tracks-dataset}.
}
\keyword{datasets}
