% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCD.r
\name{step_it}
\alias{step_it}
\title{Iterate through C-step}
\usage{
step_it(z, m, data, it = 0)
}
\arguments{
\item{z}{A vector of 0 and 1, indicating which observations should be considered for the calculation}

\item{m}{An integer specifying number of observations to use}

\item{data}{An \linkS4class{intData} object containing the macrodata/interval data}

\item{it}{An optional integer specifying the number of C-steps to perform.
With it = 0, C-step will be performed until convergence}
}
\value{
A list of z, covariance, barycenter and robust distances
}
\description{
Iterate through C-step
}
\keyword{internal}
